get-modbus/requirements.txt
psycopg2-binary==2.9.10
pyaml==25.1.0
pymodbus==3.8.6
PyYAML==6.0.2
